moved trap manager to trap folder, and enemy manager to enemy folder
